Add get_message() to FluentLocalizations.
